Obiektowa klasa Zbior
